flipperbot/requirements.txt
discord.py==1.5.1
google-cloud-pubsub==2.1.0
google-cloud-storage==1.32.0
python-dotenv==0.14.0
